Automatically discover F# files
Explicitly specifying F# files is not very developer friendly.
So I added code to automatically detect F# files in the project
directory, much the same as the C# compiler is doing.

This does mean that the compiler isn't backwards compatible with
older projects. Thing will not break, but there can be a
difference in outcome, because files are now included that weren't
included before.